Revert "Code Push"
This reverts commit 7a8d40dda3e032ca2b14f88231fd3940ffb352ff.

+++ SpirngDataJPATest/src/main/java/com/inorise/dao/FilmRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.Repository;
 
 import com.inorise.domain.Film;
 
-public interface FilmRepository extends JpaRepository<Film, Long> {
+public interface FilmRepository extends JpaRepository<Film, Short> {
                                          
 }
